requirements.txt
certifi==2020.12.5
chardet==3.0.4
idna==2.6
python-dotenv==0.17.0
requests==2.18.4
urllib3==1.22
